moves TextInputDialog.tsx from ui folder to components folder as ui folder should be used for base components while the parent component folder should have the custom built ones
