+++ vite-react/src/components/TextInputDialog.tsx
